+++ app/components/login/login-button-back.js
@@ -3,10 +3,19 @@ import { observer } from 'mobx-react/native';
 import ButtonBack from '../shared/button-back';
 import SafeComponent from '../shared/safe-component';
 import loginState from './login-state';
+import tm from '../../telemetry';
+import { telemetry } from '../../lib/icebear';
+
+const { S } = telemetry;
 
 @observer
 export default class LoginButtonBack extends SafeComponent {
+    onBackPressed() {
+        tm.login.navigate(S.BACK);
+        loginState.clearLastUser();
+    }
+
     renderThrow() {
-        return <ButtonBack onBackPressed={loginState.clearLastUser} />;
+        return <ButtonBack onBackPressed={this.onBackPressed} />;
     }
 }