+++ app/components/shared/styled-text-input.js
@@ -11,6 +11,7 @@ import icons from '../helpers/icons';
 import testLabel from '../helpers/test-label';
 import { tx } from '../utils/translator';
 import { socket } from '../../lib/icebear';
+import tm from '../../telemetry';
 
 // Because JS has no enums
 const VALID = true;
@@ -29,6 +30,7 @@ export default class StyledTextInput extends SafeComponent {
     @observable focusedAnim;
     @observable errorTextCopy;
     @observable isDirty = false;
+    prevTextLength = 0;
 
     constructor(props) {
         super(props);
@@ -62,6 +64,7 @@ export default class StyledTextInput extends SafeComponent {
     @action.bound setCustomError(error) {
         this.valid = INVALID;
         this.errorTextCopy = error;
+        tm.shared.textInputOnError(this.props.inputName, error);
     }
 
     // Checks if text field is empty and validates accordingly
@@ -76,7 +79,7 @@ export default class StyledTextInput extends SafeComponent {
                 }
                 if (required && this.isDirty) {
                     this.valid = INVALID;
-                    this.errorTextCopy = tx('title_required');
+                    this.errorTextCopy = 'error_fieldRequired';
                 }
             }
         } catch (error) {
@@ -120,6 +123,7 @@ export default class StyledTextInput extends SafeComponent {
                         this.valid = valid;
                         if (valid === INVALID) {
                             this.errorTextCopy = validation.message;
+                            tm.shared.textInputOnError(this.props.inputName, this.errorTextCopy);
                             return false;
                         }
                         return true;
@@ -134,16 +138,24 @@ export default class StyledTextInput extends SafeComponent {
         promise = promise.catch(() => {
             // Do nothing
         });
+
+        await promise;
     }
 
     @action.bound async onChangeText(text) {
         this.isDirty = true;
+        // key is entered and the key is '@'
+        if (this.props.tmTrackEmailError &&
+            this.prevTextLength + 1 === text.length && text[text.length - 1] === '@') {
+            tm.login.onLoginWithEmail(this.props.label, tx('error_usingEmailInUsernameField'));
+        }
         let inputText = text;
         const { Version, OS } = Platform;
         if (OS !== 'android' || Version > 22) {
             inputText = this.props.lowerCase ? text.toLowerCase() : text;
         }
         this.props.state.value = inputText;
+        this.prevTextLength = inputText.length;
         this.validate();
     }
 
@@ -157,10 +169,11 @@ export default class StyledTextInput extends SafeComponent {
     };
 
     @action.bound async onBlur() {
-        if (!this.props.state.value) this.validate();
         uiState.focusedTextBox = null;
         this.focused = false;
         if (this.props.onBlur) this.props.onBlur();
+        await this.validate();
+        if (this.valid === INVALID) tm.shared.textInputOnBlur(this.props.inputName, this.errorTextCopy);
     }
 
     @action.bound onFocus() {
@@ -168,6 +181,7 @@ export default class StyledTextInput extends SafeComponent {
         this.focused = true;
         if (this.props.onFocus) this.props.onFocus();
         this.textInput.focus();
+        tm.shared.textInputOnFocus(this.props.inputName);
     }
 
     get borderColor() {
@@ -196,6 +210,7 @@ export default class StyledTextInput extends SafeComponent {
         const { state } = this.props;
         // we don't give user the ability to hide passphrase again, because Apple
         this.showSecret = !this.showSecret;
+        tm.login.toggleAkVisibility(this.showSecret);
         // prevent cursor skip
         if (state.value && Platform.OS === 'android') this._skip = true;
     }
@@ -203,6 +218,7 @@ export default class StyledTextInput extends SafeComponent {
     @action.bound clearInputValue() {
         this.props.state.value = '';
         this.onChangeText('');
+        tm.shared.textInputOnClear(this.props.inputName);
     }
 
     get customIcon() {
@@ -257,10 +273,11 @@ export default class StyledTextInput extends SafeComponent {
     }
 
     get helperText() {
+        const { helperText } = this.props;
         return (
             <View style={styledTextInput.bottomMessageContainer}>
                 <Text style={styledTextInput.helperTextStyle}>
-                    {tx(this.props.helperText)}
+                    {tx(helperText)}
                 </Text>
             </View>
         );
@@ -317,6 +334,7 @@ export default class StyledTextInput extends SafeComponent {
 StyledTextInput.propTypes = {
     state: PropTypes.any,
     style: PropTypes.any,
+    inputName: PropTypes.string,
     validations: PropTypes.any,
     label: PropTypes.any,
     testID: PropTypes.any,
@@ -329,5 +347,6 @@ StyledTextInput.propTypes = {
     maxLength: PropTypes.number,
     onBlur: PropTypes.any,
     onFocus: PropTypes.any,
-    lowerCase: PropTypes.bool
+    lowerCase: PropTypes.bool,
+    tmTrackEmailError: PropTypes.bool
 };