+++ app/components/shared/top-drawer-components.js
@@ -10,10 +10,13 @@ import AvatarCircle from './avatar-circle';
 import SafeComponent from './safe-component';
 import { chatState } from '../states';
 import signupState from '../signup/signup-state';
-import { config } from '../../lib/icebear';
+import { config, telemetry } from '../../lib/icebear';
 import drawerState from './drawer-state';
 import preferenceStore from '../settings/preference-store';
 import fileState from '../files/file-state';
+import tm from '../../telemetry';
+
+const { S } = telemetry;
 
 const MAINTENANCE_DAY = 'May 15';
 const MAINTENANCE_TIME1 = '2 AM';
@@ -41,6 +44,10 @@ const innerCircle = {
 @observer
 class TopDrawerBackupAccountKey extends SafeComponent {
     renderThrow() {
+        const action = () => {
+            signupState.saveAccountKey();
+            tm.signup.saveAk(S.TERMS_TOP_DRAWER);
+        };
         return (
             <TopDrawer
                 {...this.props}
@@ -53,7 +60,7 @@ class TopDrawerBackupAccountKey extends SafeComponent {
                     </View>)}
                 descriptionLine1={tx('title_backupAkReminderMobile')}
                 buttonText={tx('button_backupNow')}
-                buttonAction={signupState.saveAccountKey}
+                buttonAction={action}
             />
         );
     }