+++ app/components/signup/signup-button-back.js
@@ -3,10 +3,19 @@ import { observer } from 'mobx-react/native';
 import signupState from './signup-state';
 import SafeComponent from '../shared/safe-component';
 import ButtonBack from '../shared/button-back';
+import tm from '../../telemetry';
+import { telemetry } from '../../lib/icebear';
+
+const { S } = telemetry;
 
 @observer
 export default class SignupButtonBack extends SafeComponent {
+    onBackPressed() {
+        tm.signup.navigate(S.BACK);
+        signupState.prev();
+    }
+
     renderThrow() {
-        return <ButtonBack onBackPressed={signupState.prev} />;
+        return <ButtonBack onBackPressed={this.onBackPressed} />;
     }
 }