+++ app/components/signup/signup-cancel.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { action } from 'mobx';
 import { observer } from 'mobx-react/native';
-import { View, Linking } from 'react-native';
+import { View } from 'react-native';
 import Text from '../controls/custom-text';
 import { vars, signupStyles } from '../../styles/styles';
 import signupState from './signup-state';
@@ -10,9 +10,14 @@ import SafeComponent from '../shared/safe-component';
 import buttons from '../helpers/buttons';
 import { TopDrawerBackupAccountKey } from '../shared/top-drawer-components';
 import { drawerState } from '../states';
-import { config, socket } from '../../lib/icebear';
+import { socket, telemetry } from '../../lib/icebear';
 import SignupHeading from './signup-heading';
 import routes from '../routes/routes';
+import TmHelper from '../../telemetry/helpers';
+import tm from '../../telemetry';
+import { popupTOS, popupPrivacy } from '../shared/popups';
+
+const { S } = telemetry;
 
 const buttonContainer = {
     flexDirection: 'row',
@@ -21,35 +26,39 @@ const buttonContainer = {
     marginTop: vars.spacing.small.mini,
     marginBottom: vars.spacing.small.maxi2x
 };
-const linkStyle = {
-    color: vars.peerioBlue
-};
 
 @observer
 export default class SignupCancel extends SafeComponent {
     componentDidMount() {
+        this.startTime = Date.now();
+        TmHelper.currentRoute = S.CANCEL_SIGN_UP;
         if (!signupState.keyBackedUp) {
             drawerState.addDrawer(TopDrawerBackupAccountKey);
         }
     }
 
-    @action.bound openPrivacyLink(text) {
-        return (
-            <Text style={linkStyle} onPress={() => { Linking.openURL(config.translator.urlMap.openPrivacy); }}>
-                {text}
-            </Text>
-        );
+    componentWillUnmount() {
+        tm.signup.duration(this.startTime);
     }
 
     @action.bound openTermsLink(text) {
-        return (
-            <Text style={linkStyle} onPress={() => { Linking.openURL(config.translator.urlMap.openTerms); }}>
-                {text}
-            </Text>
-        );
+        const onPress = async () => {
+            tm.signup.readMorePopup(S.TERMS_OF_USE);
+            await popupTOS();
+        };
+        return (<Text style={{ color: vars.peerioBlue }} onPress={onPress}>{text}</Text>);
+    }
+
+    @action.bound openPrivacyLink(text) {
+        const onPress = async () => {
+            tm.signup.readMorePopup(S.PRIVACY_POLICY);
+            await popupPrivacy();
+        };
+        return (<Text style={{ color: vars.peerioBlue }} onPress={onPress}>{text}</Text>);
     }
 
     @action.bound cancel() {
+        tm.signup.declineTos();
         drawerState.dismissAll();
         signupState.exit();
     }