+++ app/components/signup/signup-step1.js
@@ -8,15 +8,21 @@ import { vars, signupStyles } from '../../styles/styles';
 import signupState from './signup-state';
 import { tx } from '../utils/translator';
 import StyledTextInput from '../shared/styled-text-input';
-import { socket, validation } from '../../lib/icebear';
+import { socket, validation, telemetry, config } from '../../lib/icebear';
 import SafeComponent from '../shared/safe-component';
 import buttons from '../helpers/buttons';
 import SignupButtonBack from './signup-button-back';
 import SignupHeading from './signup-heading';
 import SignupStepIndicator from './signup-step-indicator';
+import TmHelper from '../../telemetry/helpers';
+import tm from '../../telemetry';
+
+const { S } = telemetry;
 
 const { validators } = validation;
 const { firstName, lastName } = validators;
+
+const MAX_NAME_LENGTH = config.user.maxNameLength;
 
 @observer
 export default class SignupStep1 extends SafeComponent {
@@ -29,6 +35,8 @@ export default class SignupStep1 extends SafeComponent {
     @action.bound onSubmitFirstName() { this.lastNameInput.onFocus(); }
 
     componentDidMount() {
+        this.startTime = Date.now();
+        TmHelper.currentRoute = S.ACCOUNT_NAME;
         // QUICK SIGNUP DEV FLAG
         if (__DEV__ && process.env.PEERIO_QUICK_SIGNUP) {
             this.firstNameInput.onChangeText(capitalize(randomWords()));
@@ -45,11 +53,16 @@ export default class SignupStep1 extends SafeComponent {
         }
     }
 
+    componentWillUnmount() {
+        tm.signup.duration(this.startTime);
+    }
+
     @action.bound async handleNextButton() {
         if (this.isNextDisabled) return;
         signupState.firstName = this.firstnameState.value;
         signupState.lastName = this.lastnameState.value;
         signupState.next();
+        tm.signup.navigate(S.NEXT);
     }
 
     get isNextDisabled() {
@@ -68,9 +81,12 @@ export default class SignupStep1 extends SafeComponent {
                         autoFocus
                         state={this.firstnameState}
                         validations={firstName}
+                        inputName={S.FIRST_NAME}
                         label={`${tx('title_firstName')}*`}
-                        helperText={tx('title_hintUsername')}
-                        maxLength={24}
+                        helperText={this.firstnameState.value.length >= MAX_NAME_LENGTH ?
+                            tx('title_characterLimitReached') :
+                            null}
+                        maxLength={MAX_NAME_LENGTH}
                         required
                         clearTextIcon
                         returnKeyType="next"
@@ -81,9 +97,12 @@ export default class SignupStep1 extends SafeComponent {
                     <StyledTextInput
                         state={this.lastnameState}
                         validations={lastName}
+                        inputName={S.LAST_NAME}
                         label={`${tx('title_lastName')}*`}
-                        helperText={tx('title_hintUsername')}
-                        maxLength={24}
+                        helperText={this.lastnameState.value.length >= MAX_NAME_LENGTH ?
+                            tx('title_characterLimitReached') :
+                            null}
+                        maxLength={MAX_NAME_LENGTH}
                         required
                         clearTextIcon
                         returnKeyType="next"