+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="3.353.0"></a>
+# [3.353.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.351.0...v3.353.0) (2018-09-17)
+
+
+
+<a name="3.351.0"></a>
+# [3.351.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.350.0...v3.351.0) (2018-09-17)
+
+
+### Bug Fixes
+
+* changed property error to errorType ([90e9064](https://github.com/PeerioTechnologies/peerio-mobile/commit/90e9064))
+* max character limit error username input ([13ece61](https://github.com/PeerioTechnologies/peerio-mobile/commit/13ece61))
+
+
+
+<a name="3.350.0"></a>
+# [3.350.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.349.0...v3.350.0) (2018-09-13)
+
+
+### Features
+
+* welcome back screen [branch ch12939] ([#309](https://github.com/PeerioTechnologies/peerio-mobile/issues/309)) ([61c6650](https://github.com/PeerioTechnologies/peerio-mobile/commit/61c6650))
+
+
+
 <a name="3.346.0"></a>
 # [3.346.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.345.0...v3.346.0) (2018-08-31)